SHTF Gear List and What You Need to Know
There are some people actually hoping for a SHTF situation. They are tired of all the needy people that constantly complain. I will admit, sometimes it seems like the world needs to hit the reset button, but we need to think about the implications of saying something like that. I’ve read estimates of up to 90 percent of the US population passing away during hypothetical SHTF scenarios. 90 percent… that’s not the criminals and bad people in the world, that’s friends and family.